JSON is right, bug fixes.
Release candidate-esque status.  Timings JSON is how it's going to be,
for now.  Offsets allow negative numbers now, which is so that if the
timings for a chord and lyric are really close, the chord will trigger
the lyric, which looks better.  Timings JSON is good enough, chord and
lyric labeled "GO" is where the song starts now.  There's still a ton of
things to fix, my coding style is amateurish, tons of global variables,
etc.  But it works.  Well enough that I'm going to start encoding my
songs for an upcoming house show.  Yaay!!!!!