Bug fixes, prep for 0.5.0 release
Fixed a bug where timings JSON from a newly-loaded song wasn't getting
loaded correctly in some instances.  It was due to the server not
necessarily loading the data into RAM before issuing the command to the
client to load it.  I think it's fixed now.

Fixed a problem where modulated chord information wasn't properly
resetting to the client.

Made it so that it remembers if you're in chord chart or Karaoke until
refresh.

I've tested it now with a bunch of songs and no reset, it seems to work
and keep working!!!